restricted delete implemented

+++ js/invoices.js
@@ -141,7 +141,10 @@ function displayInvoices(invoices) {
 
             const deleteButton = document.getElementById(`deleteInvoice-${invoice.id}`);
             deleteButton.addEventListener('click', async () => {
-                await deleteInvoice(invoice.id);
+                const confirmation = confirm(`Ar tikrai norite ištrinti sąskaitą: ${invoice.invoiceNumber} ; data:  ${invoice.invoiceDate}?`);
+                if (confirmation) {
+                    await deleteInvoice(invoice.id);
+                }
             });
 
             const updateButton = document.getElementById(`updateInvoice-${invoice.id}`);
